Fix user popover glitch in group
